pyproject.toml
[project]
name = "shorts-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.11",
    "celery[redis]>=5.5.3",
    "click>=8.2.1",
    "fastapi>=0.116.1",
    "google-genai>=1.26.0",
    "groq>=0.30.0",
    "loguru>=0.7.3",
    "moviepy>=2.2.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "scipy>=1.16.0",
    "uvicorn>=0.35.0",
]
